+++ cmd/rad/cmd/envInit.go
@@ -11,30 +11,23 @@ import (
 	"errors"
 	"fmt"
 	"net/http"
-	"sort"
 	"strings"
 
-	"github.com/Azure/azure-sdk-for-go/profiles/latest/resources/mgmt/resources"
-	"github.com/Azure/azure-sdk-for-go/profiles/preview/preview/subscription/mgmt/subscription"
 	"github.com/Azure/azure-sdk-for-go/sdk/to"
-	"github.com/Azure/go-autorest/autorest"
-	"github.com/Azure/go-autorest/autorest/azure/auth"
 	"github.com/mitchellh/mapstructure"
 	"github.com/spf13/cobra"
 	client_go "k8s.io/client-go/kubernetes"
 	runtime_client "sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/project-radius/radius/pkg/azure/clients"
 	"github.com/project-radius/radius/pkg/cli"
-	"github.com/project-radius/radius/pkg/cli/azure"
 	"github.com/project-radius/radius/pkg/cli/environments"
 	"github.com/project-radius/radius/pkg/cli/helm"
 	"github.com/project-radius/radius/pkg/cli/k3d"
 	"github.com/project-radius/radius/pkg/cli/kubernetes"
 	"github.com/project-radius/radius/pkg/cli/output"
 	"github.com/project-radius/radius/pkg/cli/prompt"
+	"github.com/project-radius/radius/pkg/cli/setup"
 	coreRpApps "github.com/project-radius/radius/pkg/corerp/api/v20220315privatepreview"
-	"github.com/project-radius/radius/pkg/handlers"
 )
 
 var envInitCmd = &cobra.Command{
@@ -47,33 +40,19 @@ func init() {
 	envCmd.AddCommand(envInitCmd)
 	envInitCmd.PersistentFlags().BoolP("interactive", "i", false, "Collect values for required command arguments through command line interface prompts")
 	envInitCmd.PersistentFlags().StringP("namespace", "n", "default", "Specify the namespace to use for the environment into which application resources are deployed")
-	envInitCmd.PersistentFlags().String("chart", "", "Specify a file path to a helm chart to install radius from")
-	envInitCmd.PersistentFlags().String("image", "", "Specify the radius controller image to use")
-	envInitCmd.PersistentFlags().String("tag", "", "Specify the radius controller tag to use")
-	envInitCmd.PersistentFlags().String("appcore-image", "", "Specify Application.Core RP image to use")
-	envInitCmd.PersistentFlags().String("appcore-tag", "", "Specify Application.Core RP image tag to use")
-}
-
-type sharedArgs struct {
-	Interactive  bool
-	Namespace    string
-	ChartPath    string
-	Image        string
-	Tag          string
-	AppCoreImage string
-	AppCoreTag   string
+	setup.RegisterPersistantChartArgs(envInitCmd)
+	setup.RegistePersistantAzureProviderArgs(envInitCmd)
 }
 
 type EnvKind int
 
 const (
-	Azure EnvKind = iota
-	Kubernetes
+	Kubernetes EnvKind = iota
 	Dev
 )
 
 func (k EnvKind) String() string {
-	return [...]string{"Azure", "Kubernetes", "Dev"}[k]
+	return [...]string{"Kubernetes", "Dev"}[k]
 }
 
 func initSelfHosted(cmd *cobra.Command, args []string, kind EnvKind) error {
@@ -83,32 +62,25 @@ func initSelfHosted(cmd *cobra.Command, args []string, kind EnvKind) error {
 		return err
 	}
 
-	sharedArgs, err := parseArgs(cmd)
+	interactive, err := cmd.Flags().GetBool("interactive")
 	if err != nil {
 		return err
 	}
 
-	ucpImage, err := cmd.Flags().GetString("ucp-image")
+	namespace, err := selectNamespace(cmd, "default", interactive)
 	if err != nil {
 		return err
 	}
 
-	ucpTag, err := cmd.Flags().GetString("ucp-tag")
+	chartArgs, err := setup.ParseChartArgs(cmd)
 	if err != nil {
 		return err
 	}
 
 	// Configure Azure provider for cloud resources if specified
-	azureProvider, err := parseAzureProviderFromArgs(cmd, sharedArgs.Interactive)
+	azureProvider, err := setup.ParseAzureProviderArgs(cmd, interactive)
 	if err != nil {
 		return err
-	}
-
-	if kind == Kubernetes {
-		sharedArgs.Namespace, err = selectNamespace(cmd, "default", sharedArgs.Interactive)
-		if err != nil {
-			return err
-		}
 	}
 
 	var defaultEnvName string
@@ -127,7 +99,7 @@ func initSelfHosted(cmd *cobra.Command, args []string, kind EnvKind) error {
 		return fmt.Errorf("unknown environment type: %s", kind)
 	}
 
-	environmentName, err := selectEnvironment(cmd, defaultEnvName, sharedArgs.Interactive)
+	environmentName, err := selectEnvironment(cmd, defaultEnvName, interactive)
 	if err != nil {
 		return err
 	}
@@ -143,15 +115,15 @@ func initSelfHosted(cmd *cobra.Command, args []string, kind EnvKind) error {
 	}
 
 	cliOptions := helm.CLIClusterOptions{
-		Namespace: sharedArgs.Namespace,
 		Radius: helm.RadiusOptions{
-			ChartPath:     sharedArgs.ChartPath,
-			Image:         sharedArgs.Image,
-			Tag:           sharedArgs.Tag,
-			UCPImage:      ucpImage,
-			UCPTag:        ucpTag,
-			AppCoreImage:  sharedArgs.AppCoreImage,
-			AppCoreTag:    sharedArgs.AppCoreTag,
+			Reinstall:     chartArgs.Reinstall,
+			ChartPath:     chartArgs.ChartPath,
+			Image:         chartArgs.Image,
+			Tag:           chartArgs.Tag,
+			UCPImage:      chartArgs.UcpImage,
+			UCPTag:        chartArgs.UcpTag,
+			AppCoreImage:  chartArgs.AppCoreImage,
+			AppCoreTag:    chartArgs.AppCoreTag,
 			AzureProvider: azureProvider,
 		},
 	}
@@ -159,9 +131,7 @@ func initSelfHosted(cmd *cobra.Command, args []string, kind EnvKind) error {
 	clusterOptions := helm.PopulateDefaultClusterOptions(cliOptions)
 
 	var k8sGoClient client_go.Interface
-	var runtimeClient runtime_client.Client
 	var contextName string
-
 	switch kind {
 	case Dev:
 		// Create environment
@@ -171,7 +141,7 @@ func initSelfHosted(cmd *cobra.Command, args []string, kind EnvKind) error {
 			return err
 		}
 		output.CompleteStep(step)
-		k8sGoClient, runtimeClient, _, err = createKubernetesClients(cluster.ContextName)
+		k8sGoClient, _, _, err = createKubernetesClients(cluster.ContextName)
 		if err != nil {
 			return err
 		}
@@ -181,7 +151,7 @@ func initSelfHosted(cmd *cobra.Command, args []string, kind EnvKind) error {
 			"kind":        "dev",
 			"context":     cluster.ContextName,
 			"clustername": cluster.ClusterName,
-			"namespace":   sharedArgs.Namespace,
+			"namespace":   namespace,
 			"registry": &environments.Registry{
 				PushEndpoint: cluster.RegistryPushEndpoint,
 				PullEndpoint: cluster.RegistryPullEndpoint,
@@ -189,22 +159,29 @@ func initSelfHosted(cmd *cobra.Command, args []string, kind EnvKind) error {
 		}
 
 	case Kubernetes:
-		k8sGoClient, runtimeClient, contextName, err = createKubernetesClients("")
+		k8sGoClient, _, contextName, err = createKubernetesClients("")
 		if err != nil {
 			return err
 		}
 		env.Items[params.Name] = map[string]interface{}{
 			"kind":      environments.KindKubernetes,
 			"context":   contextName,
-			"namespace": sharedArgs.Namespace,
+			"namespace": namespace,
 		}
-
 	}
 
-	step := output.BeginStep("Installing Radius...")
-	if err := helm.InstallOnCluster(cmd.Context(), clusterOptions, k8sGoClient, runtimeClient); err != nil {
+	// Make sure namespace for applications exists
+	err = kubernetes.EnsureNamespace(cmd.Context(), k8sGoClient, namespace)
+	if err != nil {
 		return err
 	}
+
+	err = setup.Install(cmd.Context(), clusterOptions, contextName)
+	if err != nil {
+		return err
+	}
+
+	step := output.BeginStep("Creating Environment...")
 
 	// As decided by the team we will have a temporary 1:1 correspondence between UCP resource group and environment
 	ucpRgName := fmt.Sprintf("%s-rg", environmentName)
@@ -346,185 +323,6 @@ func createKubernetesClients(contextName string) (client_go.Interface, runtime_c
 	return client, runtimeClient, contextName, nil
 }
 
-func parseArgs(cmd *cobra.Command) (sharedArgs, error) {
-	// the below function call should never errors given a default is defined
-	interactive, err := cmd.Flags().GetBool("interactive")
-	if err != nil {
-		return sharedArgs{}, err
-	}
-	namespace, err := cmd.Flags().GetString("namespace")
-	if err != nil {
-		return sharedArgs{}, err
-	}
-	chartPath, err := cmd.Flags().GetString("chart")
-	if err != nil {
-		return sharedArgs{}, err
-	}
-	image, err := cmd.Flags().GetString("image")
-	if err != nil {
-		return sharedArgs{}, err
-	}
-	tag, err := cmd.Flags().GetString("tag")
-	if err != nil {
-		return sharedArgs{}, err
-	}
-	appcoreImage, err := cmd.Flags().GetString("appcore-image")
-	if err != nil {
-		return sharedArgs{}, err
-	}
-
-	appcoreTag, err := cmd.Flags().GetString("appcore-tag")
-	if err != nil {
-		return sharedArgs{}, err
-	}
-
-	return sharedArgs{
-		Interactive:  interactive,
-		Namespace:    namespace,
-		ChartPath:    chartPath,
-		Image:        image,
-		Tag:          tag,
-		AppCoreImage: appcoreImage,
-		AppCoreTag:   appcoreTag,
-	}, nil
-}
-
-func parseAzureProviderFromArgs(cmd *cobra.Command, interactive bool) (*azure.Provider, error) {
-	if interactive {
-		return parseAzureProviderInteractive(cmd)
-	}
-	return parseAzureProviderNonInteractive(cmd)
-}
-
-func parseAzureProviderInteractive(cmd *cobra.Command) (*azure.Provider, error) {
-	authorizer, err := auth.NewAuthorizerFromCLI()
-	if err != nil {
-		return nil, err
-	}
-
-	addAzureSPN, err := prompt.ConfirmWithDefault("Add Azure provider for cloud resources [y/N]?", prompt.No)
-	if err != nil {
-		return nil, err
-	}
-	if !addAzureSPN {
-		return &azure.Provider{}, nil
-	}
-
-	subscription, err := selectSubscription(cmd.Context(), authorizer)
-	if err != nil {
-		return nil, err
-	}
-	resourceGroup, err := selectResourceGroup(cmd.Context(), authorizer, subscription)
-	if err != nil {
-		return nil, err
-	}
-
-	fmt.Printf(
-		"\nA Service Principal Name (SPN) with a corresponding role assignment and scope for your resource group is required to create Azure resources.\n\nFor example, you can create one using the following command:\n\033[36maz ad sp create-for-rbac --role Owner --scope /subscriptions/%s/resourceGroups/%s\033[0m\n\nFor more information, see: https://docs.microsoft.com/cli/azure/ad/sp?view=azure-cli-latest#az-ad-sp-create-for-rbac and https://aka.ms/azadsp-more\n\n",
-		subscription.SubscriptionID,
-		resourceGroup,
-	)
-
-	clientID, err := prompt.Text(
-		"Enter the `appId` of the service principal used to create Azure resources:",
-		prompt.UUIDv4Validator,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	clientSecret, err := prompt.Text(
-		"Enter the `password` of the service principal used to create Azure resources:",
-		prompt.EmptyValidator,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	tenantID, err := prompt.Text(
-		"Enter the `tenant` of the service principal used to create Azure resources:",
-		prompt.UUIDv4Validator,
-	)
-	if err != nil {
-		return nil, err
-	}
-
-	return &azure.Provider{
-		SubscriptionID: subscription.SubscriptionID,
-		ResourceGroup:  resourceGroup,
-		ServicePrincipal: &azure.ServicePrincipal{
-			ClientID:     clientID,
-			ClientSecret: clientSecret,
-			TenantID:     tenantID,
-		},
-	}, nil
-}
-
-func parseAzureProviderNonInteractive(cmd *cobra.Command) (*azure.Provider, error) {
-	subscriptionID, err := cmd.Flags().GetString("provider-azure-subscription")
-	if err != nil {
-		return nil, err
-	}
-	resourceGroup, err := cmd.Flags().GetString("provider-azure-resource-group")
-	if err != nil {
-		return nil, err
-	}
-
-	addAzureSPN, err := cmd.Flags().GetBool("provider-azure")
-	if err != nil {
-		return nil, err
-	}
-	if !addAzureSPN {
-		if subscriptionID == "" && resourceGroup == "" {
-			return nil, nil
-		}
-		return &azure.Provider{
-			SubscriptionID: subscriptionID,
-			ResourceGroup:  resourceGroup,
-		}, nil
-	}
-	clientID, err := cmd.Flags().GetString("provider-azure-client-id")
-	if err != nil {
-		return nil, err
-	}
-	clientSecret, err := cmd.Flags().GetString("provider-azure-client-secret")
-	if err != nil {
-		return nil, err
-	}
-	tenantID, err := cmd.Flags().GetString("provider-azure-tenant-id")
-	if err != nil {
-		return nil, err
-	}
-	if isValid, _ := prompt.UUIDv4Validator(subscriptionID); !isValid {
-		return nil, fmt.Errorf("--provider-azure-subscription is required to configure Azure provider for cloud resources")
-	}
-	if resourceGroup == "" {
-		return nil, fmt.Errorf("--provider-azure-resource-group is required to configure Azure provider for cloud resources")
-	}
-	if isValid, _ := prompt.UUIDv4Validator(clientID); !isValid {
-		return nil, errors.New("--provider-azure-client-id parameter is required to configure Azure provider for cloud resources")
-	}
-	if clientSecret == "" {
-		return nil, errors.New("--provider-azure-client-secret parameter is required to configure Azure provider for cloud resources")
-	}
-	if isValid, _ := prompt.UUIDv4Validator(tenantID); !isValid {
-		return nil, errors.New("--provider-azure-tenant-id parameter is required to configure Azure provider for cloud resources")
-	}
-	if (subscriptionID != "") != (resourceGroup != "") {
-		return nil, fmt.Errorf("to use the Azure provider both --provider-azure-subscription and --provider-azure-resource-group must be provided")
-	}
-
-	return &azure.Provider{
-		SubscriptionID: subscriptionID,
-		ResourceGroup:  resourceGroup,
-		ServicePrincipal: &azure.ServicePrincipal{
-			ClientID:     clientID,
-			ClientSecret: clientSecret,
-			TenantID:     tenantID,
-		},
-	}, nil
-}
-
 func selectNamespace(cmd *cobra.Command, defaultVal string, interactive bool) (string, error) {
 	var val string
 	var err error
@@ -563,152 +361,4 @@ func selectEnvironment(cmd *cobra.Command, defaultVal string, interactive bool) 
 		}
 	}
 	return val, nil
-}
-
-func selectSubscription(ctx context.Context, authorizer autorest.Authorizer) (azure.Subscription, error) {
-	subs, err := azure.LoadSubscriptionsFromProfile()
-	if err != nil {
-		// Failed to load subscriptions from the user profile, fall back to online.
-		subs, err = azure.LoadSubscriptionsFromAzure(ctx, authorizer)
-		if err != nil {
-			return azure.Subscription{}, err
-		}
-	}
-
-	if subs.Default != nil {
-		confirmed, err := prompt.ConfirmWithDefault(fmt.Sprintf("Use Subscription '%v'? [Y/n]", subs.Default.DisplayName), prompt.Yes)
-		if err != nil {
-			return azure.Subscription{}, err
-		}
-
-		if confirmed {
-			return *subs.Default, nil
-		}
-	}
-
-	// build prompt to select from list
-	sort.Slice(subs.Subscriptions, func(i, j int) bool {
-		l := strings.ToLower(subs.Subscriptions[i].DisplayName)
-		r := strings.ToLower(subs.Subscriptions[j].DisplayName)
-		return l < r
-	})
-	names := make([]string, 0, len(subs.Subscriptions))
-	for _, s := range subs.Subscriptions {
-		names = append(names, s.DisplayName)
-	}
-
-	index, err := prompt.SelectWithDefault("Select Subscription:", &names[0], names)
-	if err != nil {
-		return azure.Subscription{}, err
-	}
-
-	return subs.Subscriptions[index], nil
-}
-
-func selectResourceGroup(ctx context.Context, authorizer autorest.Authorizer, sub azure.Subscription) (string, error) {
-
-	rgc := clients.NewGroupsClient(sub.SubscriptionID, authorizer)
-	name, err := promptUserForRgName(ctx, rgc)
-	if err != nil {
-		return "", err
-	}
-	resp, err := rgc.CheckExistence(ctx, name)
-	if err != nil {
-		return "", err
-	} else if !resp.HasHTTPStatus(404) {
-		// already exists
-		return name, nil
-	}
-	output.LogInfo("Resource Group '%v' will be created...", name)
-
-	location, err := promptUserForLocation(ctx, authorizer, sub)
-	if err != nil {
-		return "", err
-	}
-	_, err = rgc.CreateOrUpdate(ctx, name, resources.Group{
-		Location: to.StringPtr(*location.Name),
-	})
-	if err != nil {
-		return "", err
-	}
-
-	return name, nil
-}
-
-func promptUserForLocation(ctx context.Context, authorizer autorest.Authorizer, sub azure.Subscription) (subscription.Location, error) {
-	// Use the display name for the prompt
-	// alphabetize so the list is stable and scannable
-	subc := clients.NewSubscriptionClient(authorizer)
-
-	locations, err := subc.ListLocations(ctx, sub.SubscriptionID)
-	if err != nil {
-		return subscription.Location{}, fmt.Errorf("cannot list locations: %w", err)
-	}
-
-	names := make([]string, 0, len(*locations.Value))
-	nameToLocation := map[string]subscription.Location{}
-	for _, loc := range *locations.Value {
-
-		names = append(names, *loc.DisplayName)
-		nameToLocation[*loc.DisplayName] = loc
-	}
-	sort.Strings(names)
-
-	index, err := prompt.SelectWithDefault("Select a location:", &names[0], names)
-	if err != nil {
-		return subscription.Location{}, err
-	}
-	selected := names[index]
-	return nameToLocation[selected], nil
-}
-
-func promptUserForRgName(ctx context.Context, rgc resources.GroupsClient) (string, error) {
-	var name string
-	createNewRg, err := prompt.ConfirmWithDefault("Create a new Resource Group? [Y/n]", prompt.Yes)
-	if err != nil {
-		return "", err
-	}
-	if createNewRg {
-
-		defaultRgName := "radius-rg"
-		if resp, err := rgc.CheckExistence(ctx, defaultRgName); !resp.HasHTTPStatus(404) || err != nil {
-			// only generate a random name if the default doesn't exist already or existence check fails
-			defaultRgName = handlers.GenerateRandomName("radius", "rg")
-		}
-
-		promptStr := fmt.Sprintf("Enter a Resource Group name [%s]:", defaultRgName)
-		name, err = prompt.TextWithDefault(promptStr, &defaultRgName, prompt.EmptyValidator)
-		if err != nil {
-			return "", err
-		}
-	} else {
-		rgListResp, err := rgc.List(ctx, "", nil)
-		if err != nil {
-			return "", err
-		}
-		rgList := rgListResp.Values()
-		sort.Slice(rgList, func(i, j int) bool { return strings.ToLower(*rgList[i].Name) < strings.ToLower(*rgList[j].Name) })
-		names := make([]string, 0, len(rgList))
-		for _, s := range rgList {
-			names = append(names, *s.Name)
-		}
-
-		defaultRgName, _ := azure.LoadDefaultResourceGroupFromConfig() // ignore errors resulting from being unable to read the config ini file
-		index, err := prompt.SelectWithDefault("Select ResourceGroup:", &defaultRgName, names)
-		if err != nil {
-			return "", err
-		}
-		name = *rgList[index].Name
-	}
-	return name, nil
-}
-
-// Setup flags to configure Azure provider for cloud resources
-func registerAzureProviderFlags(cmd *cobra.Command) {
-	cmd.Flags().BoolP("provider-azure", "", false, "Add Azure provider for cloud resources")
-	cmd.Flags().String("provider-azure-subscription", "", "Azure subscription for cloud resources")
-	cmd.Flags().String("provider-azure-resource-group", "", "Azure resource-group for cloud resources")
-	cmd.Flags().StringP("provider-azure-client-id", "", "", "The client id for the service principal")
-	cmd.Flags().StringP("provider-azure-client-secret", "", "", "The client secret for the service principal")
-	cmd.Flags().StringP("provider-azure-tenant-id", "", "", "The tenant id for the service principal")
 }
