+++ cmd/rad/cmd/envInitKubernetes.go
@@ -17,9 +17,6 @@ const (
 
 func init() {
 	envInitCmd.AddCommand(envInitKubernetesCmd)
-	registerAzureProviderFlags(envInitKubernetesCmd)
-	envInitKubernetesCmd.Flags().String("ucp-image", "", "Specify the UCP image to use")
-	envInitKubernetesCmd.Flags().String("ucp-tag", "", "Specify the UCP tag to use")
 }
 
 var envInitKubernetesCmd = &cobra.Command{
