+++ pkg/cli/setup/doc.go
@@ -1,0 +1,8 @@
+// ------------------------------------------------------------
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+// ------------------------------------------------------------
+
+// package setup contains support functions for programmatically installing and uninstalling Radius. These functions
+// are intended for use by the CLI and will print output.
+package setup
